// apps/console/src/lib/subscription-plan/plan-enum.ts
export enum PlanEnum {
  BASE_MODULE = 'base_module',
  COMPLIANCE_MODULE = 'compliance_module',
  ENTITY_MANAGEMENT_MODULE = 'entity_management_module',
  REGISTRY_MODULE = 'registry_module',
  TRUST_CENTER_MODULE = 'trust_center_module',
  VULNERABILITY_MANAGEMENT_MODULE = 'vulnerability_management_module',
  DOMAIN_SCANNING_ADDON = 'domain_scanning_addon',
  EXTRA_EVIDENCE_STORAGE_ADDON = 'extra_evidence_storage_addon',
  POLICY_MANAGEMENT_ADDON = 'policy_management_addon',
  RISK_MANAGEMENT_ADDON = 'risk_management_addon',
}